+++ src/main/java/devut/buzzerbidder/domain/chat/dto/response/AuctionChatMessageResponse.java
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
 
 public record AuctionChatMessageResponse(
-        @Schema(description = "메시지 타입", example = "CHAT")
+        @Schema(description = "메시지 타입", example = "AUCTION_CHAT")
         String type,
         @Schema(description = "메시지 임시 ID (타임스탬프)", example = "1735776000000")
         String tempId,
